Refactor application structure with Application and handlers separation
Move all code from main.go into well-organized modules:
- application.go: Core Application struct with certificate management and server setup
- handlers.go: All HTTP handlers and ACME protocol implementation
- main.go: Simplified entry point for application startup
- settings.go: Enhanced with buildServerURL function

This refactoring improves code organization, testability, and maintainability while preserving all existing functionality.
